Added Order Repository,Controller,Model,changed Tests

+++ Shop/Models/Category.cs
@@ -6,10 +6,9 @@ namespace Shop.Models
 {
     public class Category
     {
-
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public byte Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public virtual ICollection<Product> Products { get; set; }
     }
